% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{add}
\alias{add}
\title{Sum a list of matrices element-wise}
\usage{
add(x)
}
\arguments{
\item{x}{List of matrices of identical dimensions}
}
\value{
Single matrix that is the element-wise sum of all matrices in x
}
\description{
Sum a list of matrices element-wise
}
\examples{
mats <- list(matrix(1:4, 2, 2), matrix(1:4, 2, 2))
add(mats)

}
