% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.R
\name{al.stats}
\alias{al.stats}
\title{Compute alteration landscape statistics}
\usage{
al.stats(al)
}
\arguments{
\item{al}{SelectX object (list containing al, W, etc. as returned by selectX)}
}
\value{
ALS object with overall and per-block alteration statistics.
}
\description{
Compute alteration landscape statistics
}
