% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.R
\name{am.pairwise.alteration.overlap}
\alias{am.pairwise.alteration.overlap}
\title{Compute pairwise alteration co-occurrence counts}
\usage{
am.pairwise.alteration.overlap(am)
}
\arguments{
\item{am}{Binary alteration matrix (features x samples)}
}
\value{
Square matrix of pairwise co-occurrence counts (features x features).
}
\description{
Compute pairwise alteration co-occurrence counts
}
\examples{
am <- matrix(c(0,1,1,0,1,1), nrow = 2,
             dimnames = list(c("geneA","geneB"), c("s1","s2","s3")))
am.pairwise.alteration.overlap(am)

}
