% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{am.stats}
\alias{am.stats}
\title{Compute summary statistics for a binary alteration matrix}
\usage{
am.stats(am)
}
\arguments{
\item{am}{Binary alteration matrix (features x samples)}
}
\value{
AMS object with basic counts: n.samples, n.alterations, n.occurrences, per-sample and per-feature counts.
}
\description{
Compute summary statistics for a binary alteration matrix
}
