% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{am.weight.pairwise.alteration.overlap}
\alias{am.weight.pairwise.alteration.overlap}
\title{Compute TMB-weighted pairwise alteration overlap}
\usage{
am.weight.pairwise.alteration.overlap(am, W)
}
\arguments{
\item{am}{Binary alteration matrix (features x samples)}

\item{W}{Weight matrix (features x samples) with per-sample TMB weights}
}
\value{
Weighted pairwise overlap matrix (features x features).
}
\description{
Compute TMB-weighted pairwise alteration overlap
}
