% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.R
\name{binary.yule}
\alias{binary.yule}
\title{Compute Yule Q coefficient for all gene pairs}
\usage{
binary.yule(overlap, mat)
}
\arguments{
\item{overlap}{The pairwise overlap matrix}

\item{mat}{The binary GAM (features x samples)}
}
\value{
Matrix of Yule Q coefficients
}
\description{
Compute Yule Q coefficient for all gene pairs
}
