% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_column}
\alias{filter_maf_column}
\title{Filter maf function}
\usage{
filter_maf_column(maf, values, column, inclusive = TRUE, fixed = TRUE, ...)
}
\arguments{
\item{maf}{a maf as dataframe}

\item{values}{a list containing the elements to filter}

\item{column}{column in maf file to filter}

\item{inclusive}{a boolean to include or exclude the data frame with values in list provided}

\item{fixed}{a grep argument to specify if grep use the argument as string or not}

\item{...}{Other options}
}
\value{
filtered_maf  a filtered maf file
}
\description{
\code{filter_maf_column()} takes a maf file and filters a MAF dataframe by retaining only the rows with a column value included in the values list
}
\examples{
data(luad_maf, package = "SelectSim")
filter_maf_column(luad_maf, values = "Missense_Mutation",
                  column = "Variant_Classification")

}
