% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_complex}
\alias{filter_maf_complex}
\title{Filter a MAF dataframe by a combination of column values}
\usage{
filter_maf_complex(maf, values, ...)
}
\arguments{
\item{maf}{A MAF dataframe}

\item{values}{A dataframe of (column, value) pairs to match against}

\item{...}{Additional arguments passed to merge}
}
\value{
Filtered MAF dataframe containing only rows matching the value combinations.
}
\description{
Filter a MAF dataframe by a combination of column values
}
