% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.R
\name{filter_maf_gene.name}
\alias{filter_maf_gene.name}
\title{Filter a MAF dataframe by gene name}
\usage{
filter_maf_gene.name(maf, genes, gene.col = "Hugo_Symbol", ...)
}
\arguments{
\item{maf}{A MAF dataframe}

\item{genes}{Vector of gene names to retain}

\item{gene.col}{Column name containing gene symbols}

\item{...}{Additional arguments passed to filter_maf_column}
}
\value{
Filtered MAF dataframe.
}
\description{
Filter a MAF dataframe by gene name
}
\examples{
data(luad_maf, package = "SelectSim")
filter_maf_gene.name(luad_maf, genes = c("TP53", "KRAS"))

}
