% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_mutation.type}
\alias{filter_maf_mutation.type}
\title{Filter a MAF dataframe by mutation type}
\usage{
filter_maf_mutation.type(
  maf,
  variants,
  variant.col = "Variant_Classification",
  ...
)
}
\arguments{
\item{maf}{A MAF dataframe}

\item{variants}{Vector of variant classification values to retain}

\item{variant.col}{Column name containing variant classifications}

\item{...}{Additional arguments passed to filter_maf_column}
}
\value{
Filtered MAF dataframe.
}
\description{
Filter a MAF dataframe by mutation type
}
