% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.R
\name{filter_maf_mutations}
\alias{filter_maf_mutations}
\title{Filter a MAF dataframe by specific gene-mutation combinations}
\usage{
filter_maf_mutations(
  maf,
  values,
  maf.col = c("Hugo_Symbol", "HGVSp_Short"),
  values.col = maf.col,
  ...
)
}
\arguments{
\item{maf}{A MAF dataframe}

\item{values}{Dataframe of allowed (gene, mutation) combinations}

\item{maf.col}{Columns in maf to join on}

\item{values.col}{Corresponding columns in values to join on}

\item{...}{Additional arguments passed to filter_maf_complex}
}
\value{
Filtered MAF dataframe containing only rows matching the allowed combinations.
}
\description{
Filter a MAF dataframe by specific gene-mutation combinations
}
