% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_schema}
\alias{filter_maf_schema}
\title{This function filters a MAF dataframe by sample id}
\usage{
filter_maf_schema(maf, schema = TCGA_maf_schema, column, values, ...)
}
\arguments{
\item{maf}{a maf as dataframe}

\item{schema}{a data-frame schema; see \code{TCGA_maf_schema} for an example}

\item{column}{column in maf file to filter}

\item{values}{a list containing the elements to file}

\item{...}{Other options}
}
\value{
filtered_maf  a filtered maf file
}
\description{
\code{filter_maf_schema()} takes a maf file and filters a MAF dataframe by retaining only the rows with a column value included in the values list
}
\examples{
data(luad_maf, package = "SelectSim")
filter_maf_schema(luad_maf, TCGA_maf_schema,
                  column = "mutation.type",
                  values = TCGA_maf_schema$mutation.type$truncating)

}
