% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{GENIE_maf_schema}
\alias{GENIE_maf_schema}
\title{GENIE_maf_schema: schema for GENIE maf file to process the mutations}
\usage{
GENIE_maf_schema
}
\value{
A list defining the expected GENIE MAF column schema.
}
\description{
GENIE_maf_schema: schema for GENIE maf file to process the mutations
}
\examples{
str(GENIE_maf_schema)
}
