% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luad_maf}
\alias{luad_maf}
\title{Lung adenocarcinoma MAF from TCGA cohort}
\format{
A data frame
}
\usage{
data(luad_maf)
}
\value{
A data frame containing TCGA LUAD mutation data.
}
\description{
MAF file of LUAD from TCGA
}
\examples{
data(luad_maf)
dim(luad_maf)
head(luad_maf)
}
\keyword{datasets}
