% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luad_result}
\alias{luad_result}
\title{Lung adenocarcinoma from TCGA cohort as SelectSim run results}
\format{
A data frame
}
\usage{
data(luad_result)
}
\value{
A data frame containing SelectSim results for the LUAD cohort.
}
\description{
Result of LUAD generated by SelectSim with SelectSim run done at min.freq=10.
}
\examples{
data(luad_result)
dim(luad_result)
head(luad_result)
}
\keyword{datasets}
