% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{mutation_type}
\alias{mutation_type}
\title{Mutation list object}
\usage{
mutation_type
}
\value{
A list containing the supported mutation-type classifications.
}
\description{
Mutation list object
}
\examples{
str(mutation_type)
}
