% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{new.ALS}
\alias{new.ALS}
\title{Initialize an Alteration Landscape Stats (ALS) container}
\usage{
new.ALS(al)
}
\arguments{
\item{al}{The alteration landscape object (checked for NULL)}
}
\value{
Empty ALS list object.
}
\description{
Initialize an Alteration Landscape Stats (ALS) container
}
