% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{new.AMS}
\alias{new.AMS}
\title{Initialize an Alteration Matrix Stats (AMS) container}
\usage{
new.AMS(am)
}
\arguments{
\item{am}{The alteration matrix (checked for NULL)}
}
\value{
Empty AMS list object.
}
\description{
Initialize an Alteration Matrix Stats (AMS) container
}
\examples{
new.AMS(matrix(0, 2, 2))

}
