% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_create.R
\name{null_model_parallel}
\alias{null_model_parallel}
\title{Generating the null_simulation matrix}
\usage{
null_model_parallel(al, temp_mat, W, n.cores = 1, n.permut, seed = 42)
}
\arguments{
\item{al}{Alteration landscape object}

\item{temp_mat}{template matrices}

\item{W}{weight matrix}

\item{n.cores}{Number of cores}

\item{n.permut}{Number of simulations}

\item{seed}{Random seed}
}
\value{
List of simulated binary matrices (one per permutation)
}
\description{
Generating the null_simulation matrix
}
