% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_create.r
\name{null_model_parallel_dynamic_test}
\alias{null_model_parallel_dynamic_test}
\title{Generating the null_simulation matrix with row-specific correction and merging using pmax (experimental)}
\usage{
null_model_parallel_dynamic_test(
  al,
  temp_mat,
  W,
  n.cores = 1,
  n.permut,
  seed = 42,
  maxDiff = 0.005,
  maxIter = 100
)
}
\arguments{
\item{al}{Alteration landscape object}

\item{temp_mat}{template matrices}

\item{W}{weight matrix}

\item{n.cores}{Number of cores}

\item{n.permut}{Number of simulations}

\item{seed}{Random seed}

\item{maxDiff}{Maximum allowable error for row sums}

\item{maxIter}{Maximum number of iterations for threshold adjustment}
}
\value{
A list containing the null simulations
}
\description{
Generating the null_simulation matrix with row-specific correction and merging using pmax (experimental)
}
