% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncokb_genes}
\alias{oncokb_genes}
\title{OncoKB v3.9 cancer genes}
\format{
A list
}
\usage{
data(oncokb_genes)
}
\value{
An object containing cancer-associated genes annotated by OncoKB.
}
\description{
cancer genes list
}
\examples{
data(oncokb_genes)
length(oncokb_genes)
head(oncokb_genes)
}
\keyword{datasets}
