% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oncokb_truncating_genes}
\alias{oncokb_truncating_genes}
\title{OncoKB v3.9 cancer genes consider for truncating mutations}
\format{
A list
}
\usage{
data(oncokb_truncating_genes)
}
\value{
An object containing genes considered for truncating-mutation analyses.
}
\description{
cancer genes list consider for truncating mutations
}
\examples{
data(oncokb_truncating_genes)
length(oncokb_truncating_genes)
head(oncokb_truncating_genes)
}
\keyword{datasets}
