% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_plot.R
\name{overlap_pair_extract}
\alias{overlap_pair_extract}
\title{Extract null-model weighted overlap distribution for a gene pair}
\usage{
overlap_pair_extract(gene1, gene2, obj)
}
\arguments{
\item{gene1}{Name of the first gene/alteration.}

\item{gene2}{Name of the second gene/alteration.}

\item{obj}{SelectX object (list returned by \code{selectX()$obj}).}
}
\value{
Numeric vector of length \code{obj$nSim} with the null-model
weighted overlap values for the pair.
}
\description{
Returns the vector of weighted co-mutation values from all null-model
permutations for a specific pair of genes.
}
\examples{
\donttest{
data(luad_run_data, package = "SelectSim")
result <- selectX(M = luad_run_data$M,
                  sample.class = luad_run_data$sample.class,
                  alteration.class = luad_run_data$alteration.class,
                  n.cores = 1, min.freq = 10, n.permut = 10,
                  verbose = FALSE)
genes <- rownames(result$obj$al$am$full)[1:2]
overlap_pair_extract(genes[1], genes[2], result$obj)
}

}
