% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.R
\name{r.am.pairwise.alteration.overlap}
\alias{r.am.pairwise.alteration.overlap}
\title{Compute null overlap matrix}
\usage{
r.am.pairwise.alteration.overlap(null, n.permut, n.cores = 1)
}
\arguments{
\item{null}{The null model (list of simulated binary matrices)}

\item{n.permut}{The number of permutation steps}

\item{n.cores}{The number of cores}
}
\value{
overlap the overlap summed across permutations
}
\description{
Compute null overlap matrix
}
