% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{r.effectSize}
\alias{r.effectSize}
\title{Compute effect sizes for null model permutations}
\usage{
r.effectSize(null_overlap, mean_mat, n.permut = 1000, n.cores = 1)
}
\arguments{
\item{null_overlap}{List of null model overlap matrices (one per permutation)}

\item{mean_mat}{Mean overlap matrix across all permutations}

\item{n.permut}{Number of permutations}

\item{n.cores}{Number of cores (currently unused; sequential only)}
}
\value{
List of effect size vectors, one per permutation.
}
\description{
Compute effect sizes for null model permutations
}
