% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_run.R
\name{selectX_debug}
\alias{selectX_debug}
\title{SelectX main function to create alteration object with background model}
\usage{
selectX_debug(
  M,
  sample.class,
  alteration.class,
  n.cores = 1,
  min.freq = 10,
  n.permut = 1000,
  lambda = 0.3,
  tao = 1,
  save.object = FALSE,
  folder = "./",
  verbose = TRUE,
  estimate_pairwise = FALSE,
  maxFDR = 0.25,
  seed = 42
)
}
\arguments{
\item{M}{a list data object which consist of gams and tmbs.}

\item{sample.class}{sample covariates as named list.}

\item{alteration.class}{alteration covariates as named list.}

\item{n.cores}{no of cores.}

\item{min.freq}{number of samples for features to be atleast mutated in.}

\item{n.permut}{number of simulations.}

\item{lambda}{lambda parameter.}

\item{tao}{tao parameter.}

\item{save.object}{store the SelectX object.}

\item{folder}{folder path to store the results.}

\item{verbose}{print the time and each steps.}

\item{estimate_pairwise}{Compute pairwise p-value}

\item{maxFDR}{FDR value}

\item{seed}{a random seed}
}
\value{
result  a SelectX object with background model and other info along with result table
}
\description{
\code{selectX()} takes a list object which consist of genome alteration matrix and tumor mutation burden.
}
