% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{stat_maf_column}
\alias{stat_maf_column}
\title{Summary functions for MAF file}
\usage{
stat_maf_column(maf, column, ...)
}
\arguments{
\item{maf}{a maf as dataframe}

\item{column}{a data-frame schema; see \code{TCGA_maf_schema} for an example}

\item{...}{Other options}
}
\value{
filtered_maf  a filtered maf file
}
\description{
\code{stat_maf_column()} takes a maf file and filters a MAF dataframe by retaining only the rows with a column value included in the values list
}
\examples{
data(luad_maf, package = "SelectSim")
stat_maf_column(luad_maf, column = "Variant_Classification")

}
