% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.R
\name{stat_maf_gene}
\alias{stat_maf_gene}
\title{Count mutations per gene in a MAF file}
\usage{
stat_maf_gene(maf, column = "Hugo_Symbol", ...)
}
\arguments{
\item{maf}{A MAF dataframe}

\item{column}{Column name containing gene symbols}

\item{...}{Additional arguments passed to stat_maf_column}
}
\value{
Table of mutation counts per gene.
}
\description{
Count mutations per gene in a MAF file
}
\examples{
data(luad_maf, package = "SelectSim")
head(stat_maf_gene(luad_maf))

}
