% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.R
\name{stat_maf_sample}
\alias{stat_maf_sample}
\title{Count mutations per sample in a MAF file}
\usage{
stat_maf_sample(maf, column = "Tumor_Sample_Barcode", ...)
}
\arguments{
\item{maf}{A MAF dataframe}

\item{column}{Column name containing sample IDs}

\item{...}{Additional arguments passed to stat_maf_column}
}
\value{
Table of mutation counts per sample.
}
\description{
Count mutations per sample in a MAF file
}
