% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_create.r
\name{templeate.obj.gen}
\alias{templeate.obj.gen}
\title{Generating the template matrix}
\usage{
templeate.obj.gen(al)
}
\arguments{
\item{al}{Alteration landscape object}
}
\value{
Template matrix as list object
}
\description{
Computing the expected mutation background matrix using which simulation is done.
}
