% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_plot.R
\name{theme_Publication}
\alias{theme_Publication}
\title{A clean ggplot2 theme for publication-quality plots}
\usage{
theme_Publication(base_size = 14, base_family = "sans")
}
\arguments{
\item{base_size}{size of fonts}

\item{base_family}{type of font}
}
\value{
A ggplot2 theme object.
}
\description{
A clean ggplot2 theme for publication-quality plots
}
\examples{
library(ggplot2)
ggplot(data.frame(x = 1:3, y = 1:3), aes(x, y)) +
  geom_point() + theme_Publication()

}
