% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variant_catalogue}
\alias{variant_catalogue}
\title{OncoKB v3.9 cancer genes}
\format{
A data frame
}
\usage{
data(variant_catalogue)
}
\value{
A data frame containing cancer-gene and variant annotations.
}
\description{
A dataframe cancer genes with missense mutation annotations
}
\examples{
data(variant_catalogue)
dim(variant_catalogue)
head(variant_catalogue)
}
\keyword{datasets}
