% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{w.r.am.pairwise.alteration.overlap}
\alias{w.r.am.pairwise.alteration.overlap}
\title{Compute null weighted overlap matrix}
\usage{
w.r.am.pairwise.alteration.overlap(null, W, n.permut, n.cores = 1)
}
\arguments{
\item{null}{The null model (list of simulated binary matrices)}

\item{W}{The weight matrix (features x samples)}

\item{n.permut}{Accepted for API compatibility; currently unused internally.}

\item{n.cores}{Accepted for API compatibility; currently unused internally.}
}
\value{
Weighted overlap matrix summed across all permutations.
}
\description{
Compute null weighted overlap matrix
}
\examples{
\donttest{
data(luad_run_data, package = "SelectSim")
result <- selectX(M = luad_run_data$M,
                  sample.class = luad_run_data$sample.class,
                  alteration.class = luad_run_data$alteration.class,
                  n.cores = 1, min.freq = 10, n.permut = 10,
                  verbose = FALSE)
w.r.am.pairwise.alteration.overlap(result$obj$null, W = result$obj$W$W,
                                   n.permut = result$obj$nSim)
}

}
