% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{al.pairwise.alteration.stats}
\alias{al.pairwise.alteration.stats}
\title{Compute weight overlap stats}
\usage{
al.pairwise.alteration.stats(al, als = NULL, do.blocks = FALSE)
}
\arguments{
\item{al}{The alteration landscape object}

\item{als}{The alteration stats object}

\item{do.blocks}{blockwise comptutation}
}
\value{
overlap the weight overlap between the pairs
}
\description{
Compute weight overlap stats
}
