% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.R
\name{al.stats}
\alias{al.stats}
\title{Compute alteration landscape statistics}
\usage{
al.stats(al)
}
\arguments{
\item{al}{SelectX object (list containing al, W, etc. as returned by selectX)}
}
\value{
ALS object with overall and per-block alteration statistics.
}
\description{
Compute alteration landscape statistics
}
\examples{
\donttest{
data(luad_run_data, package = "SelectSim")
result <- selectX(M = luad_run_data$M,
                  sample.class = luad_run_data$sample.class,
                  alteration.class = luad_run_data$alteration.class,
                  n.cores = 1, min.freq = 10, n.permut = 10,
                  verbose = FALSE)
stats <- al.stats(result$obj)
}

}
