% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{al.stats}
\alias{al.stats}
\title{Create an alteration stats object}
\usage{
al.stats(al)
}
\arguments{
\item{al}{The alteration landscape object}
}
\value{
als alteration landscape stats object
}
\description{
Create an alteration stats object
}
