% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{am.pairwise.alteration.coverage}
\alias{am.pairwise.alteration.coverage}
\title{Compute pairwise alteration coverage statistics}
\usage{
am.pairwise.alteration.coverage(overlap_M, M.stats, w_overlap_M)
}
\arguments{
\item{overlap_M}{The pairwise overlap matrix}

\item{M.stats}{The alteration matrix stats (from am.stats)}

\item{w_overlap_M}{The weighted pairwise overlap matrix}
}
\value{
List with overlap and w_overlap sparse matrices.
}
\description{
Compute pairwise alteration coverage statistics
}
