% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{am.pairwise.alteration.coverage}
\alias{am.pairwise.alteration.coverage}
\title{Compute weight overlap stats}
\usage{
am.pairwise.alteration.coverage(overlap_M, M.stats, w_overlap_M)
}
\arguments{
\item{overlap_M}{The overlap matrix}

\item{M.stats}{The am matrix statis}

\item{w_overlap_M}{The weighted overlap matrix}
}
\value{
stats of the am matrix
}
\description{
Compute weight overlap stats
}
