% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{am.pairwise.alteration.overlap}
\alias{am.pairwise.alteration.overlap}
\title{Compute overlap stats}
\usage{
am.pairwise.alteration.overlap(am)
}
\arguments{
\item{am}{The alteration matrix}
}
\value{
overlap the overlap between the pairs
}
\description{
Compute overlap stats
}
