% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{am.stats}
\alias{am.stats}
\title{Create an alteration matrix stats object}
\usage{
am.stats(am)
}
\arguments{
\item{am}{The alteration matrix stat object}
}
\value{
AMS object with basic stats for the matrix
}
\description{
Create an alteration matrix stats object
}
