% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{am.weight.pairwise.alteration.overlap}
\alias{am.weight.pairwise.alteration.overlap}
\title{Compute weight overlap stats}
\usage{
am.weight.pairwise.alteration.overlap(am, W)
}
\arguments{
\item{am}{The alteration matrix}

\item{W}{The weight matrix}
}
\value{
overlap the weight overlap between the pairs
}
\description{
Compute weight overlap stats
}
