% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.R
\name{binary.yule}
\alias{binary.yule}
\title{Compute Yule Q coefficient for all gene pairs}
\usage{
binary.yule(overlap, mat)
}
\arguments{
\item{overlap}{The pairwise overlap matrix}

\item{mat}{The binary GAM (features x samples)}
}
\value{
Matrix of Yule Q coefficients
}
\description{
Compute Yule Q coefficient for all gene pairs
}
\examples{
am   <- matrix(c(0,1,1,0,1,1), nrow = 2,
               dimnames = list(c("geneA","geneB"), c("s1","s2","s3")))
ovlp <- am.pairwise.alteration.overlap(am)
binary.yule(ovlp, am)

}
