% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{binary.yule}
\alias{binary.yule}
\title{Compute yule coefficent}
\usage{
binary.yule(overlap, mat)
}
\arguments{
\item{overlap}{The overlap matrixx}

\item{mat}{The gam}
}
\value{
yule coefficent
}
\description{
Compute yule coefficent
}
