% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.R
\name{effectSize}
\alias{effectSize}
\title{Compute effect size between observed and expected overlap}
\usage{
effectSize(obs, exp)
}
\arguments{
\item{obs}{The observed overlap values}

\item{exp}{The expected (null model mean) overlap values}
}
\value{
Effect size value(s)
}
\description{
Compute effect size between observed and expected overlap
}
\examples{
effectSize(obs = 5, exp = 3)
effectSize(obs = c(5, 2, 0), exp = c(3, 3, 1))

}
