% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{effectSize}
\alias{effectSize}
\title{Compute weight overlap stats}
\usage{
effectSize(obs, exp)
}
\arguments{
\item{obs}{The observed overlap}

\item{exp}{The expected overlap}
}
\value{
effect size
}
\description{
Compute weight overlap stats
}
