% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{estimate_p_val}
\alias{estimate_p_val}
\title{Compute pairwise p-value}
\usage{
estimate_p_val(robs_co, obs.co, gene1, gene2)
}
\arguments{
\item{robs_co}{The observed values}

\item{obs.co}{The expected values aka null model}

\item{gene1}{The weight matrix}

\item{gene2}{The maxFDR to keep}
}
\value{
overlap the weight overlap between the pairs
}
\description{
Compute pairwise p-value
}
