% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{estimate_pairwise_p}
\alias{estimate_pairwise_p}
\title{Compute p-values for all gene pairs in a results table}
\usage{
estimate_pairwise_p(obs, exp, results, nSim)
}
\arguments{
\item{obs}{Observed pairwise overlap matrix}

\item{exp}{List of null model overlap matrices (one per permutation)}

\item{results}{Results data frame with SFE_1 and SFE_2 columns}

\item{nSim}{Number of permutations}
}
\value{
Vector of p-values, one per row in results.
}
\description{
Compute p-values for all gene pairs in a results table
}
\examples{
\donttest{
data(luad_run_data, package = "SelectSim")
result <- selectX(M = luad_run_data$M,
                  sample.class = luad_run_data$sample.class,
                  alteration.class = luad_run_data$alteration.class,
                  n.cores = 1, min.freq = 10, n.permut = 10,
                  verbose = FALSE, estimate_pairwise = TRUE)
head(result$result[, c("SFE_1","SFE_2","pairwise_p")])
}

}
