% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{estimate_pairwise_p}
\alias{estimate_pairwise_p}
\title{Compute pairwise p-value}
\usage{
estimate_pairwise_p(obs, exp, results, nSim)
}
\arguments{
\item{obs}{The observed values}

\item{exp}{The expected values aka null model}

\item{results}{The weight matrix}

\item{nSim}{The maxFDR to keep}
}
\value{
overlap the weight overlap between the pairs
}
\description{
Compute pairwise p-value
}
