% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{estimateFDR2}
\alias{estimateFDR2}
\title{Compute FDR}
\usage{
estimateFDR2(obs, exp, nSim, maxFDR = 0.25)
}
\arguments{
\item{obs}{The observed values}

\item{exp}{The expected values aka null model}

\item{nSim}{The weight matrix}

\item{maxFDR}{The maxFDR to keep}
}
\value{
overlap the weight overlap between the pairs
}
\description{
Compute FDR
}
