% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_complex}
\alias{filter_maf_complex}
\title{This function filters a MAF dataframe by retaining only the rows with a combination of values compatible with the values}
\usage{
filter_maf_complex(maf, values, ...)
}
\description{
This function filters a MAF dataframe by retaining only the rows with a combination of values compatible with the values
}
