% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_gene.name}
\alias{filter_maf_gene.name}
\title{This function filters a MAF dataframe by sample id}
\usage{
filter_maf_gene.name(maf, genes, gene.col = "Hugo_Symbol", ...)
}
\description{
This function filters a MAF dataframe by sample id
}
