% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_ignore}
\alias{filter_maf_ignore}
\title{This function filters a MAF dataframe by retaining (or discarding) ignore mutations}
\usage{
filter_maf_ignore(maf, schema = TCGA_maf_schema, ...)
}
\arguments{
\item{maf}{a maf as dataframe}

\item{schema}{a schema of datafrane check Select::TCGA_maf_schema for example}

\item{...}{Other options}
}
\value{
filtered_maf  a filtered maf file
}
\description{
\code{filter_maf_ignore()} takes a maf file and filters a MAF dataframe by retaining only the rows with a column value included in the values list
}
