% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_mutation.type}
\alias{filter_maf_mutation.type}
\title{This function filters a MAF dataframe by sample id}
\usage{
filter_maf_mutation.type(
  maf,
  variants,
  variant.col = "Variant_Classification",
  ...
)
}
\description{
This function filters a MAF dataframe by sample id
}
