% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_mutations}
\alias{filter_maf_mutations}
\title{This function filters a MAF dataframe by sample id}
\usage{
filter_maf_mutations(
  maf,
  values,
  maf.col = c("Hugo_Symbol", "HGVSp_Short"),
  values.col = maf.col,
  ...
)
}
\description{
This function filters a MAF dataframe by sample id
}
