% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_sample}
\alias{filter_maf_sample}
\title{Filter a MAF dataframe by sample ID}
\usage{
filter_maf_sample(maf, samples, sample.col = "Tumor_Sample_Barcode", ...)
}
\arguments{
\item{maf}{A MAF dataframe}

\item{samples}{Vector of sample IDs to retain}

\item{sample.col}{Column name containing sample IDs}

\item{...}{Additional arguments passed to filter_maf_column}
}
\value{
Filtered MAF dataframe.
}
\description{
Filter a MAF dataframe by sample ID
}
\examples{
data(luad_maf, package = "SelectSim")
ids <- unique(luad_maf$Tumor_Sample_Barcode)[1:5]
filter_maf_sample(luad_maf, samples = ids)

}
