% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\name{filter_maf_sample}
\alias{filter_maf_sample}
\title{This function filters a MAF dataframe by sample id}
\usage{
filter_maf_sample(maf, samples, sample.col = "Tumor_Sample_Barcode", ...)
}
\description{
This function filters a MAF dataframe by sample id
}
