% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_create.r
\name{generateS}
\alias{generateS}
\title{Generate S matrix}
\usage{
generateS(gam, sample.weights, upperBound = 1)
}
\arguments{
\item{gam}{the gam with genes*samples}

\item{sample.weights}{the samples weights}

\item{upperBound}{clip the values greater than 1 to keep it bounded between 0 to 1}
}
\value{
S the S matrix
}
\description{
Generate S matrix
}
