% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_create.R
\name{generateW_block}
\alias{generateW_block}
\title{Generating the weight matrix taking sample covariate}
\usage{
generateW_block(al, lambda, tau)
}
\arguments{
\item{al}{altearting landscape onject}

\item{lambda}{penalty factor parameter}

\item{tau}{fold change parameter}
}
\value{
Weight matrix
}
\description{
Generating the weight matrix taking sample covariate
}
