% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_create.r
\name{generateW_mean_tmb}
\alias{generateW_mean_tmb}
\title{Generating the weight matrix}
\usage{
generateW_mean_tmb(
  tmb,
  mean_tmb,
  ngenes,
  lambda = 0.3,
  tao = 1,
  discrete = TRUE
)
}
\arguments{
\item{tmb}{TMB dataframe}

\item{mean_tmb}{TMB dataframe}

\item{ngenes}{Number of genes}

\item{lambda}{0.1 weight factor}

\item{tao}{1 Fold change factor}

\item{discrete}{True discrete weights}
}
\value{
weight matrix (i.e vector as matrix)
}
\description{
Generating the weight matrix
}
