% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\docType{data}
\name{GENIE_maf_schema}
\alias{GENIE_maf_schema}
\title{GENIE_maf_schema: GENIE_maf_schema: schema for TCGA maf file to process the mutations}
\format{
An object of class \code{list} of length 2.
}
\usage{
GENIE_maf_schema
}
\description{
GENIE_maf_schema: GENIE_maf_schema: schema for TCGA maf file to process the mutations
}
\keyword{datasets}
