% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luad_result}
\alias{luad_result}
\title{Lung adenocarcinoma from TCGA cohort as selectX run results}
\format{
A dataframe
}
\usage{
data(luad_result)
}
\description{
Result of LUAD generated by selectX with selectX run done at min.freq=10.
}
\keyword{datasets}
