% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luad_run_data}
\alias{luad_run_data}
\title{Lung adenocarcinoma from TCGA cohort as selectX run object}
\format{
A list object with GAMs and TMBs along with sample & alteration annotations.
}
\usage{
data(luad_run_data)
}
\description{
A dataset containing the genome alteration matrix and tumor mutation burden along with sample,alteration annotation as list object
}
\keyword{datasets}
