% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\docType{data}
\name{mutation_type}
\alias{mutation_type}
\title{Mutation list object}
\format{
An object of class \code{list} of length 3.
}
\usage{
mutation_type
}
\description{
Mutation list object
}
\keyword{datasets}
