% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{new.ALS}
\alias{new.ALS}
\title{Create an alteration landscape object}
\usage{
new.ALS(al)
}
\arguments{
\item{al}{The alteration landscape}
}
\value{
AlS object.
}
\description{
Create an alteration landscape object
}
