% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{new.AMS}
\alias{new.AMS}
\title{Create an alteration matrix stats object}
\usage{
new.AMS(am)
}
\arguments{
\item{am}{The alteration matrix stats object}
}
\value{
AMS object.
}
\description{
Create an alteration matrix stats object
}
