% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_create.r
\name{null_model_parallel}
\alias{null_model_parallel}
\title{Generating the null_simulation matrix}
\usage{
null_model_parallel(al, temp_mat, W, n.cores = 1, n.permut)
}
\arguments{
\item{al}{Alteration landscape object}

\item{temp_mat}{template matrixes}

\item{W}{weight matrix}

\item{n.cores}{Number of cores}

\item{n.permut}{Number of simulation}
}
\value{
Template matrix as list object

Ordering of genes impacts the results as residual subraction is not correct.
}
\description{
Generating the null_simulation matrix
}
