% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_plot.R
\name{obs_exp_scatter}
\alias{obs_exp_scatter}
\title{Create an AL object}
\usage{
obs_exp_scatter(result, title)
}
\arguments{
\item{result}{result table of selectX obj$result}

\item{title}{title of the plot}
}
\value{
a ggplot2 object with observed vs random overlap plot.
}
\description{
Create an Alteration Landscape (AL) object which contains gams and mutation burden of samples of associated gams.
}
