% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_plot.R
\name{overlap_pair_extract}
\alias{overlap_pair_extract}
\title{Extract the backgroun distribution}
\usage{
overlap_pair_extract(gene1, gene2, obj)
}
\arguments{
\item{gene1}{gene1}

\item{gene2}{gene2}

\item{obj}{selectX obj}
}
\value{
a backgorund distribution vector.
}
\description{
Create an Alteration Landscape (AL) object which contains gams and mutation burden of samples of associated gams.
}
