% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.R
\name{r.effectSize}
\alias{r.effectSize}
\title{Compute effect sizes for null model permutations}
\usage{
r.effectSize(null_overlap, mean_mat, n.permut = 1000, n.cores = 1)
}
\arguments{
\item{null_overlap}{List of null model overlap matrices (one per permutation)}

\item{mean_mat}{Mean overlap matrix across all permutations}

\item{n.permut}{Number of permutations}

\item{n.cores}{Number of cores (currently unused; sequential only)}
}
\value{
List of effect size vectors, one per permutation.
}
\description{
Compute effect sizes for null model permutations
}
\examples{
\donttest{
data(luad_run_data, package = "SelectSim")
result  <- selectX(M = luad_run_data$M,
                   sample.class = luad_run_data$sample.class,
                   alteration.class = luad_run_data$alteration.class,
                   n.cores = 1, min.freq = 10, n.permut = 10,
                   verbose = FALSE)
null_ov  <- r.am.pairwise.alteration.overlap(result$obj$null,
                                             n.permut = result$obj$nSim)
mean_mat <- Reduce("+", null_ov) / result$obj$nSim
r.effectSize(null_ov, mean_mat, n.permut = result$obj$nSim)
}

}
