% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{r.effectSize}
\alias{r.effectSize}
\title{Compute weight overlap stats}
\usage{
r.effectSize(null_overlap, mean_mat, n.permut = 1000, n.cores = 1)
}
\arguments{
\item{null_overlap}{The null model overlap}

\item{mean_mat}{The mean effect size vector}

\item{n.permut}{The number of permutation steps}

\item{n.cores}{The number of cores}
}
\value{
overlap the weight overlap between the pairs
}
\description{
Compute weight overlap stats
}
