% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_create.r
\name{retrieveOutliers}
\alias{retrieveOutliers}
\title{Removing the Outliers}
\usage{
retrieveOutliers(obj, nSim = 1000)
}
\arguments{
\item{obj}{selectX object}

\item{nSim}{Number of simulation}
}
\value{
outliers
}
\description{
Removing the Outliers
}
