% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_plot.R
\name{ridge_plot_ed}
\alias{ridge_plot_ed}
\title{Generate a pairs background plot}
\usage{
ridge_plot_ed(result_df, obj)
}
\arguments{
\item{result_df}{result table of selectX obj$result}

\item{obj}{selectX obj}
}
\value{
a ggplot2 object with observed vs random overlap plot.
}
\description{
Create an Alteration Landscape (AL) object which contains gams and mutation burden of samples of associated gams.
}
