% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_plot.R
\name{ridge_plot_ed_compare}
\alias{ridge_plot_ed_compare}
\title{Generate a pairs background plot for two dataset comaprision}
\usage{
ridge_plot_ed_compare(result_df, obj1, obj2, name1, name2)
}
\arguments{
\item{result_df}{a common result table of selectX obj$result with dataset1_w_overlap and dataset2_w_overlap columns}

\item{obj1}{selectX obj of dataset1}

\item{obj2}{selectX obj of dataset2}

\item{name1}{dataset1 name}

\item{name2}{dataset2 name}
}
\value{
a ggplot2 object with observed vs random overlap plot.
}
\description{
Create an Alteration Landscape (AL) object which contains gams and mutation burden of samples of associated gams.
}
