% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_utils.r
\docType{data}
\name{TCGA_maf_schema}
\alias{TCGA_maf_schema}
\title{TCGA_maf_schema: schema for TCGA maf file to process the mutations}
\format{
An object of class \code{list} of length 3.
}
\usage{
TCGA_maf_schema
}
\description{
TCGA_maf_schema: schema for TCGA maf file to process the mutations
}
\keyword{datasets}
