% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_plot.R
\name{theme_Publication}
\alias{theme_Publication}
\title{A nice theme for making good plots}
\usage{
theme_Publication(base_size = 14, base_family = "sans")
}
\arguments{
\item{base_size}{size of fonts}

\item{base_family}{type of font}
}
\value{
theme object
}
\description{
A nice theme for making good plots
}
