% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variant_catalogue}
\alias{variant_catalogue}
\title{OncoKB v3.9 cancer genes}
\format{
A dataframe
}
\usage{
data(variant_catalogue)
}
\description{
A dataframe cancer genes with missense mutation annotations
}
\keyword{datasets}
