% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectX_stats.r
\name{w.r.am.pairwise.alteration.overlap}
\alias{w.r.am.pairwise.alteration.overlap}
\title{Compute weight overlap stats}
\usage{
w.r.am.pairwise.alteration.overlap(null, W, n.permut, n.cores = 1)
}
\arguments{
\item{null}{The null model}

\item{W}{The weight matrix}

\item{n.permut}{The number of permutation steps}

\item{n.cores}{The number of cores}
}
\value{
overlap the weight overlap between the pairs
}
\description{
Compute weight overlap stats
}
